requirements.txt
Flask==2.3.3
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.1